Documentation

+++ QuizyfyAPI/Controllers/UsersController.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -15,6 +12,8 @@ namespace QuizyfyAPI.Controllers
 {
     [Authorize]
     [Route("api/[controller]")]
+    [Produces("application/json")]
+    [Consumes("application/json")]
     [ApiController]
     public class UsersController : ControllerBase
     {
@@ -29,9 +28,31 @@ namespace QuizyfyAPI.Controllers
             _appSettings = appSettings.Value;
         }
 
+        /// <summary>
+        /// Authenticate user by checking if he is in database.
+        /// </summary>
+        /// <param name="model">User credentials</param>
+        /// <returns>Action Result of user model</returns>
+        /// <remarks>
+        /// Sample request (this request returns **user with token if authentication went well**)  
+        ///      
+        ///     POST /users/login
+        ///     
+        ///     {
+        ///         "username": "test",
+        ///         "password": "password"
+        ///     }    
+        ///     
+        /// </remarks> 
+        /// <response code="200">Returns user with JWT bearer token.</response>
+        /// <response code="204">User with this name was not found so return nothing.</response>
+        /// <response code="406">Request data type is not in acceptable format.</response>
         [AllowAnonymous]
-        [HttpPost("login")]
-        public async Task<ActionResult<UserModel>> Login(UserModel model)
+        [HttpPost("Login")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status406NotAcceptable)]
+        public async Task<ActionResult<UserModel>> Login(UserLoginModel model)
         {
             var user = await _repository.Authenticate(model.Username, model.Password);
 
@@ -43,11 +64,33 @@ namespace QuizyfyAPI.Controllers
             return _mapper.Map<UserModel>(user);
         }
 
+        /// <summary>
+        /// Create user with given credentials.
+        /// </summary>
+        /// <param name="model">User credentials</param>
+        /// <returns>Action Result of user model</returns>
+        /// <remarks>
+        /// Sample request (this request returns **user with token**)  
+        ///      
+        ///     POST /users/register
+        ///     
+        ///     {
+        ///         "username": "test",
+        ///         "password": "password",
+        ///         "role": "user"
+        ///     }    
+        ///     
+        /// </remarks> 
+        /// <response code="400">User is already taken or you didn't provide enough data.</response>
+        /// <response code="201">User sucessfully created. Now you can login on login action.</response>
+        /// <response code="406">Request data type is not in acceptable format.</response>
         [AllowAnonymous]
-        [HttpPost("register")]
-        public async Task<ActionResult<UserModel>> Post(UserModel model)
+        [HttpPost("Register")]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status406NotAcceptable)]
+        public async Task<ActionResult<UserModel>> Register(UserRegisterModel model)
         {
-            // map dto to entity
             var user = _mapper.Map<User>(model);
 
             if (await _repository.GetUserByUsername(model.Username) != null)
@@ -65,21 +108,55 @@ namespace QuizyfyAPI.Controllers
 
             if (await _repository.SaveChangesAsync())
             {
-                return CreatedAtAction(nameof(Login), "", model);
+                return CreatedAtAction(nameof(Login), model);
             }
 
             return BadRequest();
         }
 
-
+        /// <summary>
+        /// Updates user credentials.
+        /// </summary>
+        /// <param name="id">Id of user you want to update.</param>
+        /// <param name="model">New credentials.</param>
+        /// <returns>Action Result of user model</returns>
+        /// <remarks>
+        /// Sample request (this request returns **user**)  
+        ///      
+        ///     PUT /users/1
+        ///     
+        ///     {
+        ///         "username": "test",
+        ///         "password": "password"
+        ///     }    
+        ///     
+        /// </remarks> 
+        /// <response code="404">User with given id was not found.</response>
+        /// <response code="403">You aren't user with given id.</response>
+        /// <response code="400">User with this name already exists or you didn't provide enough data.</response>
+        /// <response code="200">Returns user with changed properties.</response>
+        /// <response code="406">Request data type is not in acceptable format.</response>
+        /// <response code="401">You aren't authenticated. Please authenticate first.</response>
         [HttpPut("{id}")]
-        public async Task<ActionResult<UserModel>> Put(int id, UserModel model)
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status406NotAcceptable)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        public async Task<ActionResult<UserModel>> Put(int id, UserRegisterModel model)
         {
             var oldUser = await _repository.GetUserById(id);
 
             if (oldUser == null)
             {
                 return NotFound($"Couldn't find user with id of {id}");
+            }
+
+
+            if (!User.IsInRole(Role.Admin) && User.Identity.Name != id.ToString())
+            {
+                return Forbid("Only admin can change other users data.");
             }
 
             if (oldUser.Username != model.Username)
@@ -109,16 +186,43 @@ namespace QuizyfyAPI.Controllers
             return BadRequest();
         }
 
-
+        /// <summary>
+        /// Deletes user with given id.
+        /// </summary>
+        /// <param name="id">Id of user you want to delete.</param>
+        /// <returns>Response Code</returns>
+        /// <remarks>
+        /// Sample request (this request returns **response code only**)  
+        ///      
+        ///     DELETE /users/1
+        ///     
+        /// </remarks> 
+        /// <response code="404">User with given id was not found.</response>
+        /// <response code="403">You aren't user with given id.</response>
+        /// <response code="400">You didn't provide enough data.</response>
+        /// <response code="200">Returns ok if user deleted.</response>
+        /// <response code="401">You aren't authenticated. Please authenticate first.</response>
         [HttpDelete("{id}")]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         public async Task<IActionResult> Delete(int id)
         {
             var oldUser = await _repository.GetUserById(id);
 
-            if(oldUser == null)
+            var user = User.Identity.Name;
+
+            if (oldUser == null)
             {
                 return NotFound();
             }
+
+            if (!User.IsInRole(Role.Admin) && oldUser.Id.ToString() != user)
+            {
+                return Forbid("Only admin can delete other users.");
+            } 
 
             _repository.Delete(oldUser);
